# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Repair Tips API
  version: 1.0.0
  description: API para gerenciamento de equipamentos, peças e suas relações de compatibilidade.
  contact:
    name: Repair Tips API
servers:
  - url: http://localhost:8080
    description: Ambiente local
tags:
  - name: Parts
    description: Operações de peças
  - name: Equipaments
    description: Operações de equipamentos
paths:
  /api/parts:
    post:
      tags: [Parts]
      operationId: createPart
      summary: Cria uma peça
      description: Cria uma peça e retorna sua representação com a URI do novo recurso no header Location.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Part'
      responses:
        '201':
          description: Peça criada
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [Parts]
      operationId: listParts
      summary: Lista peças com paginação
      description: Retorna uma página de peças ordenável por parâmetros de consulta.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Página de peças
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartPage'
  /api/parts/{id}:
    parameters:
      - $ref: '#/components/parameters/PartId'
    get:
      tags: [Parts]
      operationId: findPartById
      summary: Busca uma peça pelo identificador
      description: Retorna uma peça existente pelo seu identificador.
      responses:
        '200':
          description: Peça encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Parts]
      operationId: updatePart
      summary: Atualiza uma peça
      description: Substitui os dados de uma peça existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Part'
      responses:
        '200':
          description: Peça atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Parts]
      operationId: deletePart
      summary: Remove uma peça
      description: Remove uma peça existente.
      responses:
        '204':
          description: Peça removida
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/equipaments:
    post:
      tags: [Equipaments]
      operationId: createEquipament
      summary: Cria um equipamento
      description: Cria um equipamento e associa as peças compatíveis informadas.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipament'
      responses:
        '201':
          description: Equipamento criado
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipament'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [Equipaments]
      operationId: listEquipaments
      summary: Lista equipamentos com paginação
      description: Retorna uma página de equipamentos ordenável por parâmetros de consulta.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Página de equipamentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipamentPage'
  /api/equipaments/{id}:
    parameters:
      - $ref: '#/components/parameters/EquipamentId'
    get:
      tags: [Equipaments]
      operationId: findEquipamentById
      summary: Busca um equipamento pelo identificador
      description: Retorna um equipamento existente pelo seu identificador.
      responses:
        '200':
          description: Equipamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipament'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Equipaments]
      operationId: updateEquipament
      summary: Atualiza um equipamento
      description: Substitui os dados e as peças compatíveis de um equipamento existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipament'
      responses:
        '200':
          description: Equipamento atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipament'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Equipaments]
      operationId: deleteEquipament
      summary: Remove um equipamento
      description: Remove um equipamento existente.
      responses:
        '204':
          description: Equipamento removido
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    PartId:
      name: id
      in: path
      required: true
      description: Identificador da peça
      schema:
        type: integer
        format: int64
        example: 1
    EquipamentId:
      name: id
      in: path
      required: true
      description: Identificador do equipamento
      schema:
        type: integer
        format: int64
        example: 1
    Page:
      name: page
      in: query
      description: Número da página, começando em zero
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Quantidade de itens por página
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 20
    Sort:
      name: sort
      in: query
      description: Ordenação no formato propriedade[,asc|desc]
      schema:
        type: array
        items:
          type: string
        example: [name, desc]
      style: form
      explode: true
  headers:
    Location:
      description: URI do recurso criado
      schema:
        type: string
        format: uri
  responses:
    BadRequest:
      description: Dados inválidos
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Recurso não encontrado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    InternalServerError:
      description: Erro interno
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
    Part:
      type: object
      required: [name, number]
      properties:
        id:
          type: integer
          format: int64
          description: Identificador da peça; não informado na criação
          example: 1
          readOnly: true
        name:
          type: string
          description: Nome da peça
          example: Air filter
        number:
          type: string
          description: Número ou código da peça
          example: '12345'
    Equipament:
      type: object
      required: [name, brand, category, type, partIds]
      properties:
        id:
          type: integer
          format: int64
          description: Identificador do equipamento; não informado na criação
          example: 1
          readOnly: true
        name:
          type: string
          example: HP LaserJet Pro M404dn
        brand:
          type: string
          example: HP
        category:
          type: string
          example: PRINTER
        type:
          type: string
          example: LASER
        partIds:
          type: array
          description: Identificadores das peças compatíveis; usado na criação e atualização
          writeOnly: true
          items:
            type: integer
            format: int64
          example: [1, 2]
        parts:
          type: array
          description: Peças compatíveis; preenchido nas respostas
          readOnly: true
          items:
            $ref: '#/components/schemas/Part'
    PartPage:
      $ref: '#/components/schemas/PagePartResponse'
    EquipamentPage:
      $ref: '#/components/schemas/PageEquipamentResponse'
    PagePartResponse:
      type: object
      required: [content, number, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Part'
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageEquipamentResponse:
      type: object
      required: [content, number, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Equipament'
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: https://http.dev/problems/validation-error
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Validation failed
        instance:
          type: string
          format: uri
        timestamp:
          type: string
          format: date-time
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                example: name
              message:
                type: string
                example: Name is required